booking-platform/booking-service/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.11.0

gunicorn==20.1.0

# Database Drivers
psycopg2-binary==2.9.7
pymongo==4.3.3
redis==4.5.5

python-dotenv==1.0.0

# OSM data processing
osmium==4.0.2
numpy>=1.25.0
shapely<2.0.0
requests==2.31.0
networkx==3.4.2

pytest
pytest-flask
werkzeug==2.3.8
faker
